Refactored volunteer overview file structure
