--- src/components/Pages/VolunteerOverview/Volunteer/Routes/index.js
@@ -1,2 +1,0 @@
-import Routes from "./Routes.js";
-export default Routes;
