--- src/components/Pages/VolunteerOverview/Volunteer/index.js
@@ -1,2 +1,0 @@
-import volunteer from "./Volunteer.js";
-export default Volunteer;
