+++ src/components/Pages/Volunteers/Overview/Overview.js
@@ -2,12 +2,17 @@ import React, { useState, useEffect } from "react";
 import { connect } from "react-redux";
 
 import { getAllResponses } from "store/actions";
+import { getClassList } from "store/actions/classesActions";
+
+import { withStyles } from "@material-ui/core/styles";
+import { ListItem, ListItemText } from "@material-ui/core/";
+import { ListStyles, styles, PageContainer } from "./styles.js";
 
 import SearchCard from "components/UI/SearchCard/";
 import TrainingSeriesOverview from "components/Pages/TrainingSeries/List/Overview";
 import TrainingSeriesTab from "components/Pages/TrainingSeries/List/Tab";
-import TabNavigation from "./helpers/TabNavigation.js";
-import DesktopNavigation from "./helpers/DesktopNavigation.js";
+import TabNavigation from "./helpers/TabNavigation.js.js.js";
+import DesktopNavigation from "./helpers/DesktopNavigation.js.js.js";
 
 import {
   DoubleColumn,
@@ -16,29 +21,23 @@ import {
   DashWrapper,
   MobileNav,
   DesktopNav
-} from "./styles.js";
+} from "./styles.js.js.js";
 
 
 function Dashboard(props) {
   const [topTab, setTopTab] = useState("overview");
-  const [newResponses, setNewResponses] = useState([]);
   const {
     user_id,
     history,
-    responses,
-    getAllResponses: responsesFromProps
+    getClassList
   } = props;
 
   useEffect(() => {
-    responsesFromProps();
-    setTimeout(() => {
-      responsesFromProps();
-    }, 60 * 1000);
-  }, [responsesFromProps]);
+    getTrainingSeries();
+    getClassList();
+  }, [getTrainingSeries, getClassList]);
 
-  useEffect(() => {
-    setNewResponses(responses.filter(r => !r.seen));
-  }, [responses]);
+
 
   return (
     <DashWrapper>
@@ -66,7 +65,7 @@ function Dashboard(props) {
               <Divider />
               <SearchCard
                 user_id={user_id}
-                List={classTab}
+                List={classListTab}
                 containerTourNum="1"
                 section="Classes"
                 headerTourNum={["2", "3"]}
